+++ src/core/compress.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdlib.h>
 #include <stdint.h>
-size_t compress(const uint8_t *uncompressed, size_t uncompressedSize, uint8_t **compressed);
+uint32_t compress(const uint8_t *uncompressed, uint32_t uncompressedSize, uint8_t **compressed);
