+++ src/core/decompress.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "common/restrict.h"
 #include <stdlib.h>
 #include <stdint.h>
-size_t decompress(const uint8_t *compressed, size_t compressedSize, uint8_t **decompressed);
+uint32_t decompress(const uint8_t *compressed, uint32_t compressedSize, uint8_t **decompressed);
